Centralize configuration access (Fixes #162)
