+++ trading_bot/main.py
@@ -7,9 +7,9 @@ import signal as sig
 import sys
 from datetime import datetime, timezone
 from importlib.metadata import PackageNotFoundError, version
-from typing import Dict
 
 from trading_bot.utils.state import default_state_dir
+from trading_bot.utils.config import get_config
 
 # ? Absolute imports for package context
 from trading_bot.backtester import run_backtest
@@ -31,54 +31,6 @@ except ImportError:
 
 logger = logging.getLogger(__name__)
 
-def _deep_update(base: Dict, override: Dict) -> Dict:
-    """Recursively merge ``override`` into ``base``."""
-    for key, value in override.items():
-        if isinstance(value, dict) and isinstance(base.get(key), dict):
-            _deep_update(base[key], value)
-        else:
-            base[key] = value
-    return base
-
-
-def load_config(config_dir: str | None = None) -> Dict:
-    """Load configuration with optional local overrides.
-
-    Parameters
-    ----------
-    config_dir:
-        Directory to load config files from. Defaults to the package root.
-
-    Returns
-    -------
-    dict
-        Merged configuration dictionary.
-    """
-    base_dir = config_dir or os.path.join(os.path.dirname(os.path.dirname(__file__)))
-    config_path = os.path.join(base_dir, 'config.json')
-    try:
-        with open(config_path, 'r') as f:
-            config = json.load(f)
-    except FileNotFoundError:
-        logging.warning("config.json not found, using default values")
-        config = {
-            "symbol": "BTC/USDT",
-            "timeframe": "1m",
-            "limit": 500,
-            "sma_short": 5,
-            "sma_long": 20,
-        }
-
-    local_path = os.path.join(base_dir, 'config.local.json')
-    if os.path.exists(local_path):
-        try:
-            with open(local_path, 'r') as f:
-                local_cfg = json.load(f)
-            config = _deep_update(config, local_cfg)
-        except (OSError, json.JSONDecodeError) as e:  # noqa: BLE001
-            logging.warning(f"Failed loading config.local.json: {e}")
-
-    return config
 
 
 def parse_args():
@@ -467,7 +419,7 @@ def main():
         format='%(asctime)s [%(levelname)s] %(message)s',
         datefmt='%Y-%m-%dT%H:%M:%S%z',
     )
-    config = load_config()
+    config = get_config()
     configure_alerts(config)
     args = parse_args()
     risk_config = get_risk_config(config.get('risk'), getattr(args, 'risk_overrides', {}))
